requirements.txt
astroid==2.4.2
click==7.1.2
Flask==1.1.2
Flask-Uploads==0.2.1
Flask-WTF==0.14.3
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==7.2.0
pylint==2.5.3
pytesseract==0.3.4
six==1.15.0
toml==0.10.1
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.1
